+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.2.0] - [Helpers, part I] - 2017-03-13
+## [0.2.1] - Helpers, part II - 2017-03-13
+### Added
+- 04. Helpers/pyramid.rkt
+- 04. Helpers/pyramid-random.rkt
+- 04. Helpers/blobs.rkt
+
+## [0.2.0] - Helpers, part I - 2017-03-13
 ### Added
 - 04. Helpers/sorting.rkt